Update to ignore Fates devices for kernel updates
Does this get the job done?

+++ update/191230/update.sh
@@ -35,25 +35,29 @@ if [ -d /home/we/webdav ]; then
   sudo rm /etc/systemd/system/webdav.service
 fi
 
-# samba
-SAMBA=$(dpkg -s samba | grep "install ok")
-if [ "$SAMBA" == "" ]; then
-  sudo dpkg -i packages/*.deb
-  (echo "sleep"; echo "sleep") | sudo smbpasswd -s -a we
-  sudo cp config/smb.conf /etc/samba/
-  sudo /etc/init.d/samba restart
-fi
+# ignore kernel and samba update for Fates devices
+FATES=$(sudo cat /boot/config.txt | grep fates)
+if [[ "$FATES" == *""* ]]; then
+  # samba
+  SAMBA=$(dpkg -s samba | grep "install ok")
+  if [ "$SAMBA" == "" ]; then
+    sudo dpkg -i packages/*.deb
+    (echo "sleep"; echo "sleep") | sudo smbpasswd -s -a we
+    sudo cp config/smb.conf /etc/samba/
+    sudo /etc/init.d/samba restart
+  fi
 
-# kernel
-sudo rm /boot/kernel-*
-sudo cp -r kernel/boot /
-sudo cp -r kernel/lib /
-HW=$(sudo cat /sys/firmware/devicetree/base/model)
-if [[ "$HW" == *"Compute"* ]]; then
-  echo "CM3"
-else
-  echo "SHIELD"
-  sudo cp -r kernel/shield/* /boot/
+  # kernel
+  sudo rm /boot/kernel-*
+  sudo cp -r kernel/boot /
+  sudo cp -r kernel/lib /
+  HW=$(sudo cat /sys/firmware/devicetree/base/model)
+  if [[ "$HW" == *"Compute"* ]]; then
+    echo "CM3"
+  else
+    echo "SHIELD"
+    sudo cp -r kernel/shield/* /boot/
+  fi
 fi
 
 # logs
